--- bookmytable/src/components/css/dummy.css
@@ -1,3 +1,0 @@
-.dummy {
-  height: 50px;
-}
