--- bookmytable/src/components/css/filter.css
@@ -1,24 +1,0 @@
-.filterBox {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 12px;
-  box-sizing: border-box;
-  background: white;
-  border: 1px solid #a5a5a591;
-  border-radius: 3px;
-  box-shadow: 1px 1px 1px #a5a5a520;
-}
-
-.filterBox p {
-  margin: 0;
-}
-
-.filterBox .headings {
-  font-weight: 700;
-  font-size: 0.95rem;
-}
-
-.filterBox .filterOption {
-  /* font-weight: 700; */
-  font-size: 0.8rem;
-}
