+++ bookmytable/src/components/css/home.css
@@ -1,16 +1,6 @@
-.homeTab img {
-  width: 100%;
-  z-index: 0;
-  height: 60%;
-}
-
 #cityInput {
-  position: absolute;
-  top: 0%;
-  left: 0%;
-  width: 40%;
-  font-size: 3rem;
+  margin-top: 100px;
+  font-size: 45px;
   padding: 10px;
-  border-radius: 0px;
-  z-index: 10;
+  border-radius: 10px;
 }
