--- bookmytable/src/components/css/login.css
@@ -1,46 +1,0 @@
-.loginPage {
-  background: rgba(123, 123, 123, 0.5);
-  width: 100%;
-  height: 100%;
-  position: fixed;
-  top: 0;
-  right: 0;
-}
-
-.loginPage .loginContainer {
-  background: #fffffff3;
-  width: 100%;
-  height: 100%;
-  padding: 10px;
-}
-
-.loginPage h4 {
-  color: red;
-  margin-bottom: 20px;
-  font-weight: 900;
-}
-
-.loginPage .loginDiv {
-  padding: 10px;
-  /* border-radius: 10px; */
-  /* box-shadow: 0px 0px 2px gray; */
-  /* background: rgba(255, 205, 0, 0.1); */
-  margin: 50px 200px auto 50px;
-
-  /* box-shadow: 0px 0px 2px orange; */
-}
-
-.loginPage .loginContainer button[type="submit"] {
-  width: 100%;
-}
-
-.loginPage .loginContainer img {
-  width: 75%;
-}
-
-.loginPage .loginContainer .alert {
-  /* position: relative; */
-  widows: max-content !important;
-  font-size: 1rem;
-  padding: 2px 5px;
-}
