--- bookmytable/src/components/css/navigationBar.css
@@ -1,9 +1,0 @@
-.navBar-design {
-  background: #da0023; /* rgb(124, 20, 20);*/
-}
-
-.userBar-design {
-  background: white;
-  padding: 3px;
-  border-bottom: 1px solid #5e5e5e50;
-}
