+++ bookmytable/src/components/css/restaurants.css
@@ -3,11 +3,6 @@
   margin: 1px;
 }
 
-hr {
-  margin-bottom: 2px !important;
-  margin-top: 2px !important;
-}
-
 .restaurantCatalog {
   /* border: 2px dashed rgba(100, 1, 100, 0.4); */
   /* border-radius: 10px; */
@@ -24,12 +19,12 @@ hr {
   box-shadow: 1px 1px 1px #a5a5a520;
 }
 
-/* .restaurantDisplay:hover { */
-/* border-bottom-left-radius: 40px;
+.restaurantDisplay:hover {
+  /* border-bottom-left-radius: 40px;
   border-top-right-radius: 40px; */
-/* border: 1px solid #5555; */
-/* background: #5551; */
-/* } */
+  /* border: 1px solid #5555; */
+  /* background: #5551; */
+}
 
 .restaurantDisplay .restaurantThumbnail {
   width: 100%;
@@ -71,23 +66,11 @@ hr {
   text-align: center;
   border: none;
   background: inherit;
-  padding: 10px 20px;
+  padding: 20px;
   font-weight: 700;
   cursor: pointer;
   color: #a5a5a5;
   outline: none;
-}
-
-.activeOption {
-  color: black !important;
-}
-
-.restaurantDisplay .optionText {
-  color: red;
-  font-size: 0.8rem;
-  margin-top: 5px;
-  text-align: center;
-  width: 98%;
 }
 
 .restaurantDisplay .restaurantContacts .options:hover {
@@ -95,11 +78,6 @@ hr {
   color: black;
 }
 
-.restaurantDisplay .restaurantContacts #book:hover {
-  background: green;
-  border-radius: 5px;
-  color: white;
-}
 .restaurantDisplay .restaurantDesc p {
   margin-bottom: 2px;
 }
