+++ bookmytable/src/components/login.jsx
@@ -1,126 +1,8 @@
 import React, { Component } from "react";
-import Joi from "joi-browser";
-import "./css/login.css";
-import welcomeback from "./images/welcomeback.jpg";
 
 class Login extends Component {
-  state = { account: { email: "", password: "" }, errors: {} };
-
-  schema = {
-    email: Joi.string()
-      .email()
-      .required()
-      .label("Email"),
-    password: Joi.string()
-      .min(3)
-      .required()
-      .label("Password")
-  };
-  handleInputChange = e => {
-    let account = { ...this.state.account };
-    account[e.currentTarget.name] = e.currentTarget.value;
-    this.setState({ account });
-  };
-
-  validate = () => {
-    let errors = {};
-
-    let validationResult = Joi.validate(this.state.account, this.schema, {
-      abortEarly: false
-    });
-    if (!validationResult.error) return null;
-
-    for (let error of validationResult.error.details) {
-      errors[error.path[0]] = error.message;
-    }
-    // if (this.state.account.email.trim() === "") {
-    //   errors.email = "Email is req.";
-    // }
-    // if (this.state.account.password.trim() === "") {
-    //   errors.password = "Password is req.";
-    // }
-
-    return errors;
-  };
-
-  handleSubmit = e => {
-    e.preventDefault();
-
-    let errors = this.validate();
-    console.log(errors);
-    this.setState({ errors: errors || {} });
-
-    if (errors) return;
-    console.log("submitting LOGIN form");
-  };
-
   render() {
-    return (
-      <div
-        className="container-fluid"
-        style={{ width: "100%", height: "100%", padding: "0" }}
-      >
-        <div
-          className="row"
-          style={{ margin: "0", width: "100%", height: "100%" }}
-        >
-          <div className="col" />
-          <div className="col-5 loginContainer">
-            <button onClick={this.props.closeTab} className="btn btn-danger">
-              <i className="fa fa-times" aria-hidden="true" />
-            </button>
-            <div className="loginDiv">
-              <small class="form-text text-muted">Already a User?</small>
-              <h4>Login to continue</h4>
-              <form onSubmit={this.handleSubmit}>
-                <div className="form-group">
-                  <input
-                    value={this.state.account.email}
-                    onChange={this.handleInputChange}
-                    name="email"
-                    autoFocus
-                    className="form-control"
-                    id="exampleInputEmail1"
-                    aria-describedby="emailHelp"
-                    placeholder="Enter email"
-                  />
-                  {this.state.errors.email && (
-                    <div className="alert alert-danger">
-                      <small>{this.state.errors.email}</small>
-                    </div>
-                  )}
-                  <small id="emailHelp" className="form-text text-muted">
-                    We'll never share your email with anyone else.
-                  </small>
-                </div>
-                <div className="form-group">
-                  <input
-                    value={this.state.account.password}
-                    onChange={this.handleInputChange}
-                    type="password"
-                    name="password"
-                    className="form-control"
-                    id="exampleInputPassword1"
-                    placeholder="Password"
-                  />
-                  {this.state.errors.password && (
-                    <div>
-                      <div className="alert alert-danger">
-                        <small>{this.state.errors.password}</small>
-                      </div>
-                    </div>
-                  )}
-                </div>
-                <button type="submit" className="btn btn-danger">
-                  Submit
-                </button>
-              </form>
-            </div>
-            <img src={welcomeback} alt="happy image" />
-          </div>
-        </div>
-      </div>
-    );
+    return <h1>Login</h1>;
   }
 }
