+++ bookmytable/src/components/navigationBar.jsx
@@ -1,0 +1,40 @@
+import React, { Component } from "react";
+import { Link, NavLink } from "react-router-dom";
+class NaviagtionBar extends Component {
+  state = {};
+  render() {
+    return (
+      <nav className="navbar navbar-expand-lg navbar-dark bg-dark">
+        <Link className="navbar-brand" to="/">
+          Book My Table
+        </Link>
+        <button
+          className="navbar-toggler"
+          type="button"
+          data-toggle="collapse"
+          data-target="#navbarNavAltMarkup"
+          aria-controls="navbarNavAltMarkup"
+          aria-expanded="false"
+          aria-label="Toggle navigation"
+        >
+          <span className="navbar-toggler-icon" />
+        </button>
+        <div className="collapse navbar-collapse" id="navbarNavAltMarkup">
+          <div className="navbar-nav">
+            <NavLink className="nav-item nav-link" to="/home">
+              Home
+            </NavLink>
+            <NavLink className="nav-item nav-link" to="/restaurants">
+              Restaurant
+            </NavLink>
+            <NavLink className="nav-item nav-link" to="/login">
+              Login
+            </NavLink>
+          </div>
+        </div>
+      </nav>
+    );
+  }
+}
+
+export default NaviagtionBar;
