+++ bookmytable/src/components/restaurantTableConfig.json
@@ -1,0 +1,28 @@
+[
+  {
+    "_id": "1",
+    "name": "name",
+    "title": "Name"
+  },
+  {
+    "_id": "2",
+    "name": "cuisines",
+    "title": "Cuisines"
+  },
+  {
+    "_id": "3",
+    "name": "user_ratings",
+    "title": "Ratings"
+  },
+  {
+    "_id": "4",
+    "name": "timings",
+    "title": "Time"
+  },
+  {
+    "_id": "5",
+    "name": "average_cost_for_two",
+    "title": "Av. Cost"
+  },
+  { "_id": "6", "name": "phone_numbers", "title": "Call at" }
+]
