+++ bookmytable/src/dummy.css
@@ -1,0 +1,3 @@
+.dummy {
+  height: 100px;
+}
