--- bookmytable/src/services/authServices.js
@@ -1,7 +1,0 @@
-const getUser = () => {
-    return "null";
-}
-
-export default {
-    getUser: getUser
-}
