--- bookmytable/src/services/dump/ads.js
@@ -1,9 +1,0 @@
-const ads = [
-    { key: "1", src: "https://www.gourmetads.com/wp-content/uploads/2019/05/fast-food-ads-subway-300x600.jpg" },
-    { key: "2", src: "https://www.gourmetads.com/wp-content/uploads/2019/05/fast-food-ads-mcdonalds-300x600.jpg" },
-    { key: "3", src: "https://www.gourmetads.com/wp-content/uploads/2019/05/fast-food-ads-little-chipotle-300x600.jpg" }
-];
-
-export function getAds() {
-    return ads;
-}
