+++ mylog.md
@@ -1,77 +1,5 @@
 # MYLOG
 
-#### 2019-08-05 23:54:28
-
-on branch adjusting-js
-
-- login form validation corrected
-- signup form design done
-- validation for the signup done
-- authServices.js initialized
-
-#### 2019-08-05 20:15:52
-
-- login form validated
-- connect to server left
-
-#### 2019-08-05 18:09:52
-
-- Everything mergerd back to app-front
-- Filter designing done
-- login designing done
-- ads added
-- TODOS:
-  - implement JS functionality to the buttons
-  - booking table attribute
-
-#### 2019-08-05 14:45:40
-
-- Target: User and Filter
-- Integrate the backend if possible
-
-#### 2019-08-04 00:21:49
-
-- Filter added for design
-- Ads added
-- JQuery for bootstrap elements added
-- userbar added
-- login page on hold
-- home page design thought
-- branches created today:
-  - implementing-filter (child to restaurent model; filter designin)
-  - restaurant-model (child to bookTable; refactoring the restaurants page)
-  - bookTable (child to app-front; created to implement the Book a Table logic on hold)
-  - login-form (added login form and sign up form basic design not implented totally)
-
-#### 2019-08-03 18:52:24
-
-- Restaurant as seprate module refactoring complete
-- bookTable attribute is underprocess
-- Menu / Call functions done
-- App's common div refactored:
-  - container made component dependent
-- User from initiated
-- Home page and Navigation Bar refactoring begun (but on hold)
-
-#### 2019-08-02 18:19:49
-
-- Navigation bar refactoring started
-- Project On Hold due to:
-  - login forms and submit attributes
-  - user module
-  - restaurant actions implemantation
-- TODOS:
-  - refactor Home page
-  - build user
-  - build login
-  - restaurant actions
-  - **rest from #2019-08-02 01:58:33**
-
-#### 2019-08-02 15:33:08
-
-- Restaurants display refactored
-- adjusted all the attributes of the json file into the page
-
 #### 2019-08-02 01:58:33
 
 - Restaurants are being refactored to div from tables
